% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CoastGreen.R
\name{Bscore}
\alias{Bscore}
\title{Extract pixellated blue value/hex color code in an image}
\usage{
Bscore(img, plot.blue = FALSE)
}
\arguments{
\item{img}{cimg array object}

\item{plot.blue}{logical, plot blue area only.}
}
\description{
Extract pixellated blue value/hex color code in an image
}
