% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CoastGreen.R
\name{Gscore}
\alias{Gscore}
\title{Extract pixellated green value/hex color code in an image}
\usage{
Gscore(img, output = c("Garea", "HEX"), plot.green = FALSE)
}
\arguments{
\item{img}{cimg array object}

\item{output}{"Garea", coveragre (%) of green area in a picture. "HEX", hex color code of green area.}

\item{plot.green}{logical, plot green area only.}
}
\description{
Extract pixellated green value/hex color code in an image
}
