% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CoastGreen.R
\name{pixeliz}
\alias{pixeliz}
\title{Pixellate images}
\usage{
pixeliz(
  imgfile,
  rf = 0.5,
  rp = 10,
  rW = NULL,
  rH = NULL,
  resize.out = FALSE,
  plot.px = FALSE
)
}
\arguments{
\item{imgfile}{Import color image file.}

\item{rf}{Scale of resize, default is 0.5.}

\item{rp}{Pixellate image per square of rp pixels, default is 10.}

\item{rW}{Resize image to specific width. default is NULL.}

\item{rH}{Resize image to specific height default is NULL.}

\item{resize.out}{logical, default is FALSE. Output resized image}

\item{plot.px}{logical, default is FALSE. Plot pixellated image}
}
\value{
cimg format array
}
\description{
Resize and pixellate images
}
