% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CoastGreen.R
\name{repix}
\alias{repix}
\title{Extract pixellated colors}
\usage{
repix(img, rW, rH, rp)
}
\arguments{
\item{img}{cimg array object}

\item{rW}{Width of pixellated image.}

\item{rH}{Height of pixellated image.}

\item{rp}{Square of rp pixels of the pixellated image.}
}
\value{
minimized cimg format array
}
\description{
Extract representative colors from pixellated images
}
