% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CoastGreen.R
\name{Gscore}
\alias{Gscore}
\title{Extract pixellated green value/hex color code in an image}
\usage{
Gscore(img, output = c("Garea", "HEX"), plot.green = FALSE)
}
\description{
Extract pixellated green value/hex color code in an image
}
